Backend

+++ backend/src/resolvers/Mutation.js
@@ -62,7 +62,7 @@ const Mutation = {
   async deleteUser(parent, { username }, { db, request }, info) {
     console.log("resolvers/Mutation/deleteUser");
     if (!request.user) throw new Error("Unauthenticated operation");
-    if (request.user.name !== "admin")
+    if (request.user.username !== "admin")
       throw new Error("Only admin can create user");
 
     const user = await db.UserModel.findOne({
@@ -283,6 +283,10 @@ const Mutation = {
    */
   async createOpenday(parent, { data: args }, { db }, info) {
     console.log("resolvers/Mutation/createOpenday");
+    if (!request.user) throw new Error("Unauthenticated operation");
+    if (request.user.username !== "admin")
+      throw new Error("Only admin can create / update Openday");
+
     // weekday: type of string
     const existing = await db.OpendayModel.findOne({ weekday: args.weekday });
 
@@ -290,7 +294,7 @@ const Mutation = {
     if (existing)
       return await db.OpendayModel.findOneAndUpdate(
         { weekday: args.weekday },
-        { ...data },
+        { ...args },
         { new: true } // return updated
       );
 
@@ -298,7 +302,11 @@ const Mutation = {
     const openday = await new db.OpendayModel({ ...args }).save();
     return openday;
   },
-  async deleteOpenday(parent, { weekday }, { db, pubsub }, info) {
+  async deleteOpenday(parent, { weekday }, { db, request }, info) {
+    console.log("resolvers/Mutation/deleteOpenday");
+    if (!request.user) throw new Error("Unauthenticated operation");
+    if (request.user.username !== "admin")
+      throw new Error("Only admin can delete Openday");
     const deletedOpenday = await db.OpendayModel.findOneAndDelete({
       weekday: weekday,
     });
